requirements.txt
Django==2.0
django-bootstrap3-datetimepicker==2.2.3
django-datetime-widget==0.9.3
pip==9.0.1
pytz==2017.3
setuptools==28.8.0
